Hard code version in version.py
Using setuptools_scm to extract version info from git tags is a bit
awkward and means that the package can only be built from a full
standalone git repository. Of various options, defining the version in
a separate version.py file seems clean and reliable.

It remains to be seen how much we may miss the automatic development
version bumps for each commit...