Expose restraint potential objects.
For simulations run in the local context, this commit allows access to
the potentials used in the context after the simulation has run and while
the context is still alive. This allows the Python script to interact
with the object implementing the potential with any additional interface
it may have.

This is a temporary form of access until other means of interacting with
workflow objects are devised and implemented.