Clarify embedded proof-of-concept code.
Commented out some dead code more completely, but have not removed it
entirely because it captures some ideas that are still being explorered,
and there doesn't seem to be a good way to attach sample source code to
GitHub issues. References to the appropriate tracked issue have been
added.

Maybe this is something that "Gist" is good for.