gmxapi-158 Add "end_time" keyword argument for gmx.workflow.from_tpr
Update for 0.0.6 release. Add "end_time" keyword argument to from_tpr()
function and clarify behavior of `end_time` and `steps`. For 0.0.6,
`end_time` is processed by the `load_tpr` operation in `gmx.context`.
The TPR file specified as the input source for the operation is copied
and updated by the framework in a temporary area as an implementation
detail. The original TPR file is left untouched.

For the more rigorous solution in 0.0.7, refer to issue #158.