Load GmxMdParams from TPR file.
Use some additional helpers to map C++ types to GROMACS types.

As a design decision (at least for the moment), allow that all API input
will be converted to a reduced set of types (e.g. float64 and int64),
with precision reduced if needed based on the TPR file implementation.
In a future change, we could check that the provided values will fit in
the actual type when setting. Note that the basic Python API has a
single integer type and a single floating point type.