Basic user-level TPR reader.
Fixes #192

Add gmx.fileio.read_tpr() to get an object providing simulation run
input data from a TPR file. In this change, the interface to TprFile
contents is not firm or well-documented, but the functionality is
sufficient to allow proper inspection of the effects of the `end_time`
parameter provided through the gmx.core.copy_tpr function (also not
part of the public interface yet).