Pass MDWorkSpec to System::launch instead of using member.
There was ambiguity in where and whether a MDWorkSpec should be created
or held. Put responsibility for MDWorkSpec management on the client.
Create a temporary empty MDWorkSpec if necessary to allow `launch` to be
used without a `spec` argument, when extensions are not needed in the
simulation. Remove the `spec_` member from System::Impl

This should resolve a bug in which restraint plugins were not properly
attached to the MdRunner at simulation launch when using the Python
interface.

Future work should separate the `launch` functionality to be clearly the
creation routine for a SimulationSession, and the System class can be
evolved to be more clearly a SimulationInput object, only. The
MDWorkSpec should be merged into the details of the gmx::MDModules
composite, which should be passed with the Context when launching a
SimulationSession.